% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{cel2RMA}
\alias{cel2RMA}
\title{Create an RMA-normalized expression data frame from .cel files}
\usage{
cel2RMA(cel_dir = "./", pattern = ".cel")
}
\arguments{
\item{cel_dir}{Path to directory containing binary .cel files. Default is the current working directory.}

\item{pattern}{Character indicating pattern to match binary .cel files. Default is "cel", indicating that all .cel files must have "cel" in their file names.}
}
\value{
An RMA-normalized expression data frame with probe IDs as row names and arrays in column names.
}
\description{
Create an RMA-normalized expression data frame from .cel files
}
\seealso{
\code{\link[affy]{read.affybatch}},\code{\link[affy]{rma}}
 \code{\link[Biobase]{eSet}},\code{\link[Biobase]{exprs}}
}
