% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sft}
\alias{check_sft}
\title{Check scale-free topology fit for a given network}
\usage{
check_sft(edgelist, net_type = "gcn")
}
\arguments{
\item{edgelist}{Edge list as a 2-column data frame containing node 1, node 2 and edge weight.}

\item{net_type}{Type of biological network. One of "gcn", "grn", or "ppi". Default: gcn.}
}
\description{
Check scale-free topology fit for a given network
}
