% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{consensus_modules}
\alias{consensus_modules}
\title{Identify consensus modules across multiple data sets}
\usage{
consensus_modules(
  exp_list,
  setLabels = NULL,
  metadata,
  cor_method = "spearman",
  net_type = "signed hybrid",
  rsquared = 0.8,
  module_merging_threshold = 0.2
)
}
\arguments{
\item{exp_list}{A list containing the expression data frames. For each expression data frame, row names correspond to gene IDs and column names correspond to sample names. The list can be created by using \code{list(exp1, exp2, ..., expn)}.}

\item{setLabels}{Character vector containing labels for each expression set.}

\item{metadata}{Data frame containing sample information. The first column must contain sample names and the second column must contain sample information.}

\item{cor_method}{Correlation method used for network reconstruction. One of "spearman" (default), "biweight", or "pearson".}

\item{net_type}{Network type. One of "signed hybrid" (default), "signed" or "unsigned".}

\item{rsquared}{Minimum R squared to consider the network similar to a scale-free topology. Default is 0.8.}

\item{module_merging_threshold}{Cut-off for merging similar modules based on the dissimilarity of their eigengenes. Default is 0.2, indicating that modules that have correlation >= 0.8 will be merged into one.}
}
\value{
A list containing 4 elements: \describe{
  \item{consModules}{Consensus module assignments}
  \item{consMEs}{Consensus module eigengenes}
  \item{exprSize}{Description of the multi-set object returned by the function \code{WGCNA::checkSets}}
  \item{sampleInfo}{Metadata for each expression set}
}
}
\description{
This function identifies modules that are conserved across multiple data sets. It can be used for evolutionary analyses and for testing robustness of coexpression modules in data sets from different sources, for instance.
}
\seealso{
\code{\link[WGCNA]{checkSets}},\code{\link[WGCNA]{pickSoftThreshold}},\code{\link[WGCNA]{adjacency}},\code{\link[WGCNA]{TOMsimilarity}},\code{\link[WGCNA]{pquantile}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{multiSetMEs}},\code{\link[WGCNA]{consensusMEDissimilarity}},\code{\link[WGCNA]{mergeCloseModules}},\code{\link[WGCNA]{plotDendroAndColors}}
 \code{\link[pals]{discrete}}
 \code{\link[dynamicTreeCut]{cutreeDynamic}}
}
