% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{consensus_set_relationship}
\alias{consensus_set_relationship}
\title{Relate consensus modules to set-specific modules}
\usage{
consensus_set_relationship(setMEs, setColors, consMEs, consColors)
}
\arguments{
\item{setMEs}{Module eigengenes for the set-specific network. It is the 2nd element of the result list from \code{exp2net}.}

\item{setColors}{Module assignment for the set-specific network. It is the 5th element of the result list from \code{exp2net}.}

\item{consMEs}{Module eigengenes for consensus network. It is the 2nd element of the result list from \code{consensus_modules}.}

\item{consColors}{Module assignment for the consensus network. It is the 1st element of the result list from \code{consensus_modules}.}
}
\value{
Heatmap of relationships between consensus modules and set-specific modules in a PDF file.
}
\description{
Relate consensus modules to set-specific modules
}
\seealso{
\code{\link[WGCNA]{orderMEs}},\code{\link[WGCNA]{labeledHeatmap}},\code{\link[WGCNA]{blueWhiteRed}}
}
