% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{consensusmodules_sample_cor}
\alias{consensusmodules_sample_cor}
\title{Correlate set-specific modules and consensus modules to sample information}
\usage{
consensusmodules_sample_cor(
  consMEs,
  exprSize,
  sampleInfo,
  cor_method = "spearman"
)
}
\arguments{
\item{consMEs}{Data frame of consensus module eigengenes returned by \code{consensus_modules}.}

\item{exprSize}{Object containing information on the multi-set used for inference of consensus modules. Third object of the result list from \code{consensus_modules}.}

\item{sampleInfo}{Data frame of two columns containing sample names in the first column and sample information in the second column.}

\item{cor_method}{Correlation method to be used. One of 'spearman' or 'pearson'. Default is 'spearman'.}
}
\value{
Heatmaps of relationships between consensus/set-specific modules and sample information (metadata) in a PDF file.
}
\description{
Correlate set-specific modules and consensus modules to sample information
}
\seealso{
\code{\link[WGCNA]{corPvalueFisher}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{labeledHeatmap}},\code{\link[WGCNA]{blueWhiteRed}}
}
