% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R, R/network_stats.R
\name{exp2net}
\alias{exp2net}
\title{Reconstruct gene coexpression network from gene expression data frame}
\usage{
exp2net(
  norm.exp,
  net_type = "signed hybrid",
  module_merging_threshold = 0.2,
  SFTpower,
  cor_method = "spearman"
)

exp2net(
  norm.exp,
  net_type = "signed hybrid",
  module_merging_threshold = 0.2,
  SFTpower,
  cor_method = "spearman"
)
}
\arguments{
\item{norm.exp}{Normalized gene expression data frame, where row names are gene IDs and column names are sample names.}

\item{net_type}{Network type. One of 'signed', 'signed hybrid' or 'unsigned'.}

\item{module_merging_threshold}{Maximum dissimilarity (1-cor) between module for module merging. As it is based on dissimilarity, a threshold of 0.2 means that module eigengenes must be at least 0.8 similar to be merged. Default is 0.2.}

\item{SFTpower}{SFT power generated by the function \code{SFT_fit}.}

\item{cor_method}{Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman", considering that the expression data does not follow a normal distribution.}
}
\value{
List containing: \itemize{
  \item Adjacency matrix
  \item Data frame of module eigengenes
  \item Data frame of genes and their corresponding modules
  \item Data frame of intramodular connectivity
  \item Vector of module assignment
  \item Correlation matrix
  \item Parameters used for network reconstruction
}

List containing: \itemize{
  \item Adjacency matrix
  \item Data frame of module eigengenes
  \item Data frame of genes and their corresponding modules
  \item Data frame of intramodular connectivity
  \item Vector of module assignment
  \item Correlation matrix
  \item Parameters used for network reconstruction
}
}
\description{
Reconstruct gene coexpression network from gene expression data frame

Reconstruct gene coexpression network from gene expression data frame
}
\seealso{
\code{\link[WGCNA]{adjacency.fromSimilarity}},\code{\link[WGCNA]{TOMsimilarity}},\code{\link[WGCNA]{standardColors}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{moduleEigengenes}},\code{\link[WGCNA]{plotEigengeneNetworks}},\code{\link[WGCNA]{mergeCloseModules}},\code{\link[WGCNA]{plotDendroAndColors}},\code{\link[WGCNA]{intramodularConnectivity}}
 \code{\link[dynamicTreeCut]{cutreeDynamicTree}}

\code{\link[WGCNA]{adjacency.fromSimilarity}},\code{\link[WGCNA]{TOMsimilarity}},\code{\link[WGCNA]{standardColors}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{moduleEigengenes}},\code{\link[WGCNA]{plotEigengeneNetworks}},\code{\link[WGCNA]{mergeCloseModules}},\code{\link[WGCNA]{plotDendroAndColors}},\code{\link[WGCNA]{intramodularConnectivity}}
 \code{\link[dynamicTreeCut]{cutreeDynamicTree}}
}
\author{
Fabricio Almeida-Silva

Fabricio Almeida-Silva
}
