% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filt.se}
\alias{filt.se}
\title{Filtered soybean seed gene expression data from SRA study PRJNA197251}
\format{
An object of class \code{SummarizedExperiment}
}
\usage{
filt.se
}
\description{
Filtered expression data in transcripts per million (TPM) from Danzer et al., 2015.
This is the same data set described in \code{se.seed}, but it only contains the top 500 genes with the highest variances.
This data set was created to be used in unit tests and examples.
}
\examples{
data(filt.se)
}
\references{
Danzer, J., Mellott, E., Bui, A. Q., Le, B. H., Martin, P., Hashimoto, M., ... & Goldberg, R. B. (2015).
Down-regulating the expression of 53 soybean transcription factor genes uncovers a role for SPEECHLESS in
initiating stomatal cell lineages during embryo development. Plant Physiology, 168(3), 1025-1035.
}
\keyword{datasets}
