% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{filter_by_variance}
\alias{filter_by_variance}
\title{Filter expression data frame to keep only the most varying genes}
\usage{
filter_by_variance(exp, n = NULL, percentile = NULL)
}
\arguments{
\item{exp}{Expression data frame with genes as rows and samples as columns}

\item{n}{Number of most variable genes (e.g. n=5000 will keep the top 5000 most variable genes).}

\item{percentile}{Percentile of most highly variable genes (e.g. percentile=0.1 will keep the top 10 percent most variable genes). Values must range from 0 to 1.}
}
\value{
Expressed data frame with the most variable genes in row names and samples in column names.
}
\description{
Filter expression data frame to keep only the most varying genes
}
\author{
Fabricio Almeida-Silva
}
