% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{filter_edges}
\alias{filter_edges}
\title{Filter edge list to remove spurious correlations}
\usage{
filter_edges(
  cor_matrix,
  method = "optimalSFT",
  r_optimal_test = seq(0.7, 0.9, by = 0.05),
  Zcutoff = 1.96,
  pvalue_cutoff = 0.05,
  rcutoff = 0.7,
  nSamples = NULL,
  check_SFT = FALSE
)
}
\arguments{
\item{cor_matrix}{Correlation matrix, such as the one returned by \code{exp2net}. It is the last element of the resulting list from \code{exp2net}.}

\item{method}{Method to filter spurious correlations. One of "Zscore", "optimalSFT", "pvalue" or "min_cor". See details for more information on the methods. Default: 'optimalSFT'}

\item{r_optimal_test}{Numeric vector with the correlation thresholds to be tested for optimal scale-free topology fit. Only valid if \code{method} equals "optimalSFT". Default: seq(0.7, 0.9, by = 0.05)}

\item{Zcutoff}{Minimum Z-score threshold. Only valid if \code{method} equals "Zscore". Default: 1.96}

\item{pvalue_cutoff}{Maximum P-value threshold. Only valid if \code{method} equals "pvalue". Default: 0.05}

\item{rcutoff}{Minimum correlation threshold. Only valid if \code{method} equals "min_cor". Default: 0.7}

\item{nSamples}{Number of samples in the dataset from which the correlation matrix was calculated. Only required if \code{method} equals "pvalue".}

\item{check_SFT}{Logical indicating whether to test if the resulting network is scale-free or not. Default: FALSE}
}
\value{
A filtered edge list.
}
\description{
Filter edge list to remove spurious correlations
}
\details{
The default method ("optimalSFT") will create several different edge lists by filtering the original correlation matrix by the thresholds specified in \code{r_optimal_test}. Then, it will calculate a scale-free topology fit index for each of the possible networks and return the network that best fits the scale-free topology.
The method "Zscore" will apply a Fisher Z-transformation for the correlation coefficients and remove the Z-scores below the threshold specified in \code{Zcutoff}.
The method "pvalue" will calculate Student asymptotic p-value for the correlations and remove correlations whose p-values are above the threshold specified in \code{pvalue_cutoff}.
The method "min_cor" will remove correlations below the minimum correlation threshold specified in \code{rcutoff}.
}
\seealso{
\code{\link[WGCNA]{scaleFreeFitIndex}},\code{\link[WGCNA]{corPvalueStudent}}
 \code{\link[igraph]{fit_power_law}}
}
