% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{get_HK}
\alias{get_HK}
\title{Get a vector object with housekeeping genes}
\usage{
get_HK(exp)
}
\arguments{
\item{exp}{Expression data frame, where row names are gene IDs and col names are sample names}
}
\value{
Character vector of housekeeping gene IDs
}
\description{
Get a vector object with housekeeping genes
}
\author{
Fabricio Almeida-Silva
}
