% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_analysis.R
\name{get_homologs}
\alias{get_homologs}
\title{Get homologous genes of given gene or gene set}
\usage{
get_homologs(genes, dup_pairs)
}
\arguments{
\item{genes}{Character vector containing gene or set of genes of which homologs will be extracted.}

\item{dup_pairs}{Data frame containing all ortholog pairs (different species) or paralog pairs (same species). Column 1 and column 2 must contain gene 1 and gene 2 from the pair, respectively. Addiotional columns are interpreted as pairs info (e.g. mode of duplication for paralog pairs).}
}
\value{
List containing homologs of each gene in the input gene set.
}
\description{
Get homologous genes of given gene or gene set
}
\author{
Fabricio Almeida-Silva
}
