% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{get_hubs}
\alias{get_hubs}
\title{Get hub genes in a given module or in all modules at once}
\usage{
get_hubs(exp, net)
}
\arguments{
\item{exp}{Gene expression data frame with gene IDs as row names and sample names as column names.}

\item{net}{List object returned by \code{exp2net}.}
}
\value{
Data frame containing gene IDs, modules and intramodular connectivity of all hubs.
}
\description{
Get hub genes in a given module or in all modules at once
}
\seealso{
\code{\link[WGCNA]{signedKME}}
}
\author{
Fabricio Almeida-Silva
}
