% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{get_hubs_gcn}
\alias{get_hubs_gcn}
\title{Get hub genes in a given module or in all modules at once}
\usage{
get_hubs_gcn(exp, net)
}
\arguments{
\item{exp}{A gene expression data frame with genes in row names and samples in column names or a `SummarizedExperiment` object.}

\item{net}{List object returned by \code{exp2gcn}.}
}
\value{
Data frame containing gene IDs, modules and intramodular connectivity of all hubs.
}
\description{
Get hub genes in a given module or in all modules at once
}
\examples{
data(se.seed)
filt.se <- filter_by_variance(se.seed, n=500)
gcn <- exp2gcn(filt.se, SFTpower = 16, cor_method = "pearson", reportPDF = FALSE)
hubs <- get_hubs_gcn(filt.se, gcn)
}
\seealso{
\code{\link[WGCNA]{signedKME}}
}
\author{
Fabricio Almeida-Silva
}
