% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_inference.R
\name{get_hubs_grn}
\alias{get_hubs_grn}
\title{Get hubs for gene regulatory network}
\usage{
get_hubs_grn(edgelist, top_percentile = 0.1, top_n = NULL)
}
\arguments{
\item{edgelist}{A gene regulatory network represented as an edge list.}

\item{top_percentile}{Numeric from 0 to 1 indicating the percentage of genes with the highest degree to consider hubs. Default: 0.1.}

\item{top_n}{Numeric indicating the number of genes with the highest degree to consider hubs.}
}
\value{
A data frame with gene ID in the first column and out degree in the second column.
}
\description{
Get hubs for gene regulatory network
}
\examples{
data(filt.se)
tfs <- sample(rownames(filt.se), size=50, replace=FALSE)
grn_list <- grn_combined(filt.se, regulators=tfs, nTrees=2)
ranked_grn <- grn_average_rank(grn_list)
# split in only 2 groups for demonstration purposes
filtered_edges <- grn_filter(ranked_grn, nsplit=2)
hubs <- get_hubs_grn(filtered_edges)
}
