% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_inference.R
\name{grn_aracne}
\alias{grn_aracne}
\title{Infer gene regulatory network with the ARACNE algorithm}
\usage{
grn_aracne(exp)
}
\arguments{
\item{exp}{Expression matrix with gene IDs as row names and samples as column names.}
}
\value{
A gene regulatory network represented as an edge list.
}
\description{
Infer gene regulatory network with the ARACNE algorithm
}
