% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_inference.R
\name{grn_genie3}
\alias{grn_genie3}
\title{Infer gene regulatory network with GENIE3}
\usage{
grn_genie3(exp, regulators = NULL, ...)
}
\arguments{
\item{exp}{Expression matrix with gene IDs as row names and samples as column names.}

\item{regulators}{A character vector of regulators (e.g., transcription factors or miRNAs). All regulators must be included in `exp`.}

\item{...}{Additional arguments passed to `GENIE3::GENIE3()`.}
}
\value{
A gene regulatory network represented as an edge list.
}
\description{
Infer gene regulatory network with GENIE3
}
