% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stats.R, R/network_visualization.R
\name{igraph2ggnetwork}
\alias{igraph2ggnetwork}
\title{Create a ggnetwork data frame from an igraph object}
\usage{
igraph2ggnetwork(graph, layout = "kk", arrow.gap = 0.2)

igraph2ggnetwork(graph, layout = "kk", arrow.gap = 0.2)
}
\arguments{
\item{graph}{Object of class igraph.}

\item{layout}{Network layout. One of "dh", "drl", "gem", "lgl", "fr", "graphopt", "kk" and "mds". Default is "kk".}

\item{arrow.gap}{Numeric indicating the distance between nodes and arrows. Default is 0.2}
}
\description{
Create a ggnetwork data frame from an igraph object

Create a ggnetwork data frame from an igraph object
}
\seealso{
\code{\link[ggnetwork]{ggnetwork}}
 \code{\link[igraph]{layout_with_dh}},\code{\link[igraph]{layout_with_drl}},\code{\link[igraph]{layout_with_gem}},\code{\link[igraph]{layout_with_lgl}},\code{\link[igraph]{layout_with_fr}},\code{\link[igraph]{layout_with_graphopt}},\code{\link[igraph]{layout_with_kk}},\code{\link[igraph]{layout_with_mds}}

\code{\link[ggnetwork]{ggnetwork}}
 \code{\link[igraph]{layout_with_dh}},\code{\link[igraph]{layout_with_drl}},\code{\link[igraph]{layout_with_gem}},\code{\link[igraph]{layout_with_lgl}},\code{\link[igraph]{layout_with_fr}},\code{\link[igraph]{layout_with_graphopt}},\code{\link[igraph]{layout_with_kk}},\code{\link[igraph]{layout_with_mds}}
}
\author{
Fabricio Almeida-Silva

Fabricio Almeida-Silva
}
