% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_analysis.R
\name{is_duplicated}
\alias{is_duplicated}
\title{Boolean expression to check if gene or gene set is duplicated in the genome or not}
\usage{
is_duplicated(genes, dup_pairs)
}
\arguments{
\item{genes}{Character containing gene or group of genes to be evaluated.}

\item{dup_pairs}{Data frame containing all duplicate pairs. Column 1 and column 2 must contain gene 1 and gene 2 from the pair, respectively. Addiotional columns are interpreted as pairs info (e.g. mode of duplication).}
}
\value{
Vector of logical values indicating if gene or group of genes has copies in the genome or not.
}
\description{
For a given gene or group of genes, this function returns logical values indicating whether they have copies in the genome or not.
}
\seealso{
\code{is_singleton}
}
\author{
Fabricio Almeida-Silva
}
