% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolutionary_analysis.R
\name{is_singleton}
\alias{is_singleton}
\title{Boolean expression to check if gene or gene set is singleton or not}
\usage{
is_singleton(genes, dup_pairs)
}
\arguments{
\item{genes}{Character containing gene or group of genes to be evaluated.}

\item{dup_pairs}{Data frame containing all duplicate pairs. Column 1 and column 2 must contain gene 1 and gene 2 from the pair, respectively. Addiotional columns are interpreted as pairs info (e.g. mode of duplication).}
}
\value{
Vector of logical values indicating if gene or group of genes is singleton or not.
}
\description{
For a given gene or group of genes, this function returns logical values indicating whether they are singletons or not.
}
\seealso{
\code{is_duplicated}
}
\author{
Fabricio Almeida-Silva
}
