% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{modPres_netrep}
\alias{modPres_netrep}
\title{Calculate module preservation between two expression data sets using NetRep's algorithm}
\usage{
modPres_netrep(
  explist,
  ref_net = NULL,
  test_net = NULL,
  nPerm = 1000,
  nThreads = 2
)
}
\arguments{
\item{explist}{List of expression matrices with gene/probe names corresponding to column names and sample names corresponding to row names.}

\item{ref_net}{Reference network object returned by the function \code{exp2net}.}

\item{test_net}{Test network object returned by the function \code{exp2net}.}

\item{nPerm}{Number of permutations. Default: 1000}

\item{nThreads}{Number of threads to be used for parallel computing. Default: 2}
}
\value{
Output list from \code{NetRep::modulePreservation} and a message in user's standard output stating which modules are preserved.
}
\description{
Calculate module preservation between two expression data sets using NetRep's algorithm
}
\seealso{
\code{\link[NetRep]{modulePreservation}}
}
