% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{module_enrichment}
\alias{module_enrichment}
\title{Perform enrichment analysis for coexpression network modules}
\usage{
module_enrichment(
  net = NULL,
  background_genes,
  annotation,
  column = NULL,
  correction = "BH",
  p = 0.05,
  min_setsize = 10,
  max_setsize = 500,
  bp_param = BiocParallel::SerialParam()
)
}
\arguments{
\item{net}{List object returned by \code{exp2gcn}.}

\item{background_genes}{Character vector of genes to be used as background
for the Fisher's Exact Test.}

\item{annotation}{Annotation data frame with genes in the first column
and functional annotation in the other columns. This data frame can
be exported from Biomart or similar databases.}

\item{column}{Column or columns of \code{annotation} to be used for enrichment.
Both character or numeric values with column indices can be used.
If users want to supply more than one column, input a character or
numeric vector. Default: all columns from \code{annotation}.}

\item{correction}{Multiple testing correction method. One of "holm",
"hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none".
Default is "BH".}

\item{p}{P-value threshold. P-values below this threshold will be considered
significant. Default is 0.05.}

\item{min_setsize}{Numeric indicating the minimum gene set size to be
considered. Gene sets correspond to levels of each variable
in \strong{annotation}). Default: 10.}

\item{max_setsize}{Numeric indicating the maximum gene set size to be
considered. Gene sets correspond to levels of each variable
in \strong{annotation}). Default: 500.}

\item{bp_param}{BiocParallel back-end to be used.
Default: BiocParallel::SerialParam()}
}
\value{
A data frame of overrepresentation results with the following
variables:
\describe{
  \item{term}{character, functional term ID/name.}
  \item{genes}{numeric, intersection length between input genes and genes
               in a particular functional term.}
  \item{all}{numeric, number of all genes in a particular functional term.}
  \item{pval}{numeric, P-value for the hypergeometric test.}
  \item{padj}{numeric, P-value adjusted for multiple comparisons using
              the method specified in parameter \strong{adj}.}
  \item{category}{character, name of the grouping variable (i.e., column
                  name of \strong{annotation}).}
  \item{module}{character, module name.}
}
}
\description{
Perform enrichment analysis for coexpression network modules
}
\examples{
\donttest{
data(filt.se)
data(zma.interpro)
background <- rownames(filt.se)
gcn <- exp2gcn(filt.se, SFTpower = 18, cor_method = "pearson")
mod_enrich <- module_enrichment(gcn, background, zma.interpro, p=1)
}
}
\author{
Fabricio Almeida-Silva
}
