% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{module_preservation}
\alias{module_preservation}
\title{Calculate module preservation between two expression data sets}
\usage{
module_preservation(
  exp_ref,
  exp_test,
  moduleColors,
  savePreservation = TRUE,
  plot_all_stats = FALSE
)
}
\arguments{
\item{exp_ref}{Data frame of reference gene expression data set, with gene IDs in row names and sample names in column names.}

\item{exp_test}{Data frame of test gene expression data set, with gene IDs in row names and sample names in column names.}

\item{moduleColors}{Vector of module assignment returned by \code{exp2net}.}

\item{savePreservation}{Logical indicating whether to save module preservation into an R object or not. As the calculation of module preservation can take a long time, it is useful to save time in future analyses. Default is TRUE.}

\item{plot_all_stats}{Logical indicating whether to save all density and connectivity statistics in a PDF file or not. Default is FALSE.}
}
\value{
Module preservation statistics.
}
\description{
Calculate module preservation between two expression data sets
}
\seealso{
\code{\link[WGCNA]{modulePreservation}}
}
