% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{module_sample_cor}
\alias{module_sample_cor}
\title{Correlate module eigengenes to sample categories (e.g. tissues)}
\usage{
module_sample_cor(exp, metadata, MEs, cor_method = "spearman")
}
\arguments{
\item{exp}{Gene expression data frame used to reconstruct the network, with gene IDs as row names and sample names as column names.}

\item{metadata}{A 2-column data frame containing sample names in the first column and sample descriptions in the second column.}

\item{MEs}{Module eigengenes. It is the 2nd element of the result list generated by the function \code{exp2net}.}

\item{cor_method}{Method to calculate correlation. One of 'pearson', 'spearman' or 'kendall'. Default is 'spearman'.}
}
\value{
A heatmap showing the correlations between module eigengenes and sample categories with their associated P-values.
}
\description{
Correlate module eigengenes to sample categories (e.g. tissues)
}
\seealso{
\code{\link[WGCNA]{corPvalueStudent}},\code{\link[WGCNA]{labeledHeatmap}},\code{\link[WGCNA]{blueWhiteRed}}
}
\author{
Fabricio Almeida-Silva
}
