% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R, R/network_stats.R
\name{module_stability}
\alias{module_stability}
\title{Perform module stability analysis}
\usage{
module_stability(norm.exp, net, nRuns = 30)

module_stability(norm.exp, net, nRuns = 30)
}
\arguments{
\item{norm.exp}{Normalized gene expression data frame, where row names are gene IDs and column names are sample names.}

\item{net}{List object returned by \code{exp2net}.}

\item{nRuns}{Number of times to resample. Default is 30.}
}
\value{
A .pdf file in the current working directory representing the module stability across the resamplings.

A .pdf file in the current working directory representing the module stability across the resamplings.
}
\description{
Perform module stability analysis

Perform module stability analysis
}
\seealso{
\code{\link[WGCNA]{sampledBlockwiseModules}},\code{\link[WGCNA]{matchLabels}},\code{\link[WGCNA]{plotDendroAndColors}}

\code{\link[WGCNA]{sampledBlockwiseModules}},\code{\link[WGCNA]{matchLabels}},\code{\link[WGCNA]{plotDendroAndColors}}
}
