% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{og.zma.osa}
\alias{og.zma.osa}
\title{Orthogroups between maize and rice}
\format{
A 3-column data frame with orthogroups, species IDs and gene IDs.
}
\usage{
data(og.zma.osa)
}
\description{
The orthogroups were downloaded from the PLAZA 4.0 Monocots database.
}
\examples{
data(og.zma.osa)
}
\references{
Van Bel, M., Diels, T., Vancaester, E., Kreft, L., Botzki, A.,
Van de Peer, Y., ... & Vandepoele, K. (2018). PLAZA 4.0: an integrative
resource for functional, evolutionary and comparative plant
genomics. Nucleic acids research, 46(D1), D1190-D1196.
}
\keyword{datasets}
