% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{parse_orthofinder}
\alias{parse_orthofinder}
\title{Parse orthogroups identified by OrthoFinder}
\usage{
parse_orthofinder(file_path = NULL)
}
\arguments{
\item{file_path}{Path to Orthogroups/Orthogroups.tsv file generated
by OrthoFinder.}
}
\value{
A 3-column data frame with orthogroups, species IDs and
gene IDs, respectively.
}
\description{
This function converts the orthogroups file named \strong{Orthogroups.tsv} to
a 3-column data frame that can be interpreted by BioNERO.
}
\examples{
path <- system.file("extdata", "Orthogroups.tsv", package = "BioNERO")
og <- parse_orthofinder(path)
}
\author{
Fabricio Almeida-Silva
}
