% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{PC_correction}
\alias{PC_correction}
\title{Apply Principal Component (PC)-based correction for confounding artifacts in coexpression networks}
\usage{
PC_correction(exp_df)
}
\arguments{
\item{exp_df}{Expression dataframe where rownames are gene IDs and colnames are sample names. Expression data must not have undergone normalization.}
}
\value{
Corrected expression data frame
}
\description{
Apply Principal Component (PC)-based correction for confounding artifacts in coexpression networks
}
\seealso{
\code{\link[sva]{num.sv}},\code{\link[sva]{sva_network}}
}
\author{
Fabricio Almeida-Silva
}
