% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{plot_dendro_and_colors}
\alias{plot_dendro_and_colors}
\title{Plot dendrogram of genes and modules}
\usage{
plot_dendro_and_colors(gcn)
}
\arguments{
\item{gcn}{List object returned by \code{exp2gcn}.}
}
\value{
A base plot with the gene dendrogram and modules.
}
\description{
Plot dendrogram of genes and modules
}
\examples{
data(filt.se)
gcn <- exp2gcn(filt.se, SFTpower = 18, cor_method = "pearson")
plot_dendro_and_colors(gcn)
}
