% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{plot_dendro_and_cons_colors}
\alias{plot_dendro_and_cons_colors}
\title{Plot dendrogram of genes and consensus modules}
\usage{
plot_dendro_and_cons_colors(consensus)
}
\arguments{
\item{consensus}{Consensus network returned by \code{consensus_modules}.}
}
\value{
A base plot with the gene dendrogram and modules.
}
\description{
Plot dendrogram of genes and consensus modules
}
\examples{
set.seed(12)
data(zma.se)
filt.zma <- filter_by_variance(zma.se, n=500)
zma.set1 <- filt.zma[, sample(colnames(filt.zma), size=20, replace=FALSE)]
zma.set2 <- filt.zma[, sample(colnames(filt.zma), size=20, replace=FALSE)]
list.sets <- list(zma.set1, zma.set2)
# SFT power previously identified with consensus_SFT_fit()
cons_mod <- consensus_modules(list.sets, power = c(11, 13),
                              cor_method = "pearson")
plot_dendro_and_cons_colors(cons_mod)
}
