% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{plot_eigengene_network}
\alias{plot_eigengene_network}
\title{Plot eigengene network}
\usage{
plot_eigengene_network(gcn)
}
\arguments{
\item{gcn}{List object returned by \code{exp2gcn}.}
}
\value{
A base plot with the eigengene network
}
\description{
Plot eigengene network
}
\examples{
data(filt.se)
gcn <- exp2gcn(filt.se, SFTpower = 18, cor_method = "pearson")
plot_eigengene_network(gcn)
}
