% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_grn}
\alias{plot_grn}
\title{Plot gene regulatory network from edge list}
\usage{
plot_grn(
  edgelist_grn,
  show_labels = "tophubs",
  top_n_hubs = 5,
  interactive = FALSE,
  layout = "kk",
  arrow.gap = 0.01,
  ranked = TRUE
)
}
\arguments{
\item{edgelist_grn}{Data frame containing the edge list for the GRN network. First column is the TF and second column is the target gene. All other columns are interpreted as edge attributes.}

\item{show_labels}{Character indicating which nodes will be labeled. One of "all", "allhubs", "tophubs", or "none".}

\item{top_n_hubs}{Number of top hubs to be labeled. It is only valid if \code{show_labels} equals "tophubs". Default is 5.}

\item{interactive}{Logical indicating whether the network should be interactive or not. Default is FALSE.}

\item{layout}{Network layout. One of "dh", "drl", "gem", "lgl", "fr", "graphopt", "kk" and "mds". Default is "kk".}

\item{arrow.gap}{Numeric indicating the distance between nodes and arrows. Default is 0.2.}

\item{ranked}{Logical indicating whether to treat third column of the edgelist (edge weights) as ranked values. Default: TRUE.}
}
\value{
A ggplot object containing the network.
}
\description{
Plot gene regulatory network from edge list
}
\examples{
data(filt.se)
tfs <- sample(rownames(filt.se), size=50, replace=FALSE)
grn_edges <- grn_clr(filt.se, regulators = tfs)
p <- plot_grn(grn_edges, ranked=FALSE)
}
\seealso{
\code{\link[networkD3]{igraph_to_networkD3}},\code{\link[networkD3]{forceNetwork}}
 \code{\link[ggnetwork]{geom_edges}}, \code{\link[ggnetwork]{geom_nodes}},\code{\link[ggnetwork]{geom_nodetext}},\code{\link[ggnetwork]{theme_blank}},\code{\link[ggnetwork]{geom_nodetext_repel}}
 \code{\link[ggnewscale]{new_scale}}
}
\author{
Fabricio Almeida-Silva
}
