% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{plot_network}
\alias{plot_network}
\title{Plot network from edge list}
\usage{
plot_network(
  bionetwork = "GCN",
  edgelist_gcn = NULL,
  edgelist_int = NULL,
  genes_modules = NULL,
  modulename = NULL,
  kIN = NULL,
  hubs = NULL,
  top_n_hubs = 5,
  hubs_only = FALSE,
  layout = layout_with_drl(),
  facet_mod = TRUE,
  clustering_method = "infomap"
)
}
\arguments{
\item{bionetwork}{Type of biological network. One of 'GCN' (gene coexpression network), 'PPI' (protein-protein interaction), 'GRN' (gene-regulatory network), or 'Mixed' (GCN + PPI). Default is 'GCN'.}

\item{edgelist_gcn}{Data frame containing the edge list for the gene coexpression network. First column is gene 1 and second column is gene 2, which is coexpressed with gene 1. All other columns are interpreted as edge attributes (e.g. weight). The edge list can be obtained from an adjacency matrix with the function \code{get_edge_list}. If bionetwork is 'PPI' or 'GRN', this parameter can be ignored.}

\item{edgelist_int}{Data frame containing the edge list for the PPI or GRN network. First column is the protein 1 (or TF) and second column is the protein 2 (or target gene). All other columns are interpreted as edge attributes. If bionetwork is 'GCN', this parameter can be ignored.}

\item{genes_modules}{Data frame containing genes in column 1 and their corresponding modules in column 2. It is the third element of the output list from \code{exp2net}.}

\item{modulename}{Character with name of the module to be plotted. To include 2 or more modules, input the names in a character vector.}

\item{kIN}{Data frame contaning intramodular connectivity for each gene in the network. It is the fourth element of the output list from \code{exp2net}.}

\item{hubs}{Data frame containing hub genes in the first column, their modules in the second column, and intramodular connectivity in the third column.}

\item{top_n_hubs}{Number of top hubs to be labeled. Default is 5.}

\item{hubs_only}{Logical indicating if only hub genes should be plotted. Default is FALSE.}

\item{clustering_method}{Community detection algorithm to be used. Available methods are "infomap", "edge_betweenness", "fast_greedy", "walktrap", "spinglass", "leading_eigen", "louvain", and "label_prop". Default is "infomap".}

\item{facet}{Logical. If more than 1 module is chosen, should different modules be plotted in different facets? Default is TRUE.}
}
\value{
A ggplot object with the network.
}
\description{
Plot network from edge list
}
\seealso{
\code{get_edge_list}

\code{exp2net}.

\code{get_hubs}

\code{\link[igraph]{as_data_frame}},\code{\link[igraph]{degree}}
 \code{\link[ggnetwork]{ggnetwork}}
}
\author{
Fabricio Almeida-Silva
}
