% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_ngenes_per_module}
\alias{plot_ngenes_per_module}
\title{Plot number of genes per module}
\usage{
plot_ngenes_per_module(genes_and_modules = NULL)
}
\arguments{
\item{genes_and_modules}{A 2-column data frame with gene IDs in the first column and module names in the second column. Third element of the result list from \code{exp2net}.}
}
\value{
A ggplot object with a barplot of gene number in each module.
}
\description{
Plot number of genes per module
}
\seealso{
\code{\link[ggpubr]{ggbarplot}}
}
