% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{plot_ppi}
\alias{plot_ppi}
\title{Plot protein-protein interaction network from edge list}
\usage{
plot_ppi(
  edgelist_int,
  detect_communities = TRUE,
  clustering_method = "infomap",
  show_labels = "tophubs",
  top_n_hubs = 5,
  interactive = FALSE
)
}
\arguments{
\item{edgelist_int}{Data frame containing the edge list for the PPI network. First column is the protein 1 and second column is the protein 2. All other columns are interpreted as edge attributes.}

\item{detect_communities}{Logical indicating whether to detect communities or not. Default is TRUE.}

\item{clustering_method}{Community detection algorithm to be used. Available methods are "infomap", "edge_betweenness", "fast_greedy", "walktrap", "spinglass", "leading_eigen", "louvain", and "label_prop". Default is "infomap".}

\item{show_labels}{Character indicating which nodes will be labeled. One of "all", "allhubs", "tophubs", or "none".}

\item{top_n_hubs}{Number of top hubs to be labeled. It is only valid if \code{show_labels} equals "tophubs". Default is 5.}

\item{interactive}{Logical indicating whether the network should be interactive or not. Default is FALSE.}
}
\description{
Plot protein-protein interaction network from edge list
}
\seealso{
\code{\link[igraph]{as_data_frame}},\code{\link[igraph]{degree}},\code{\link[igraph]{simplify}},\code{\link[igraph]{gorder}}
 \code{\link[networkD3]{igraph_to_networkD3}},\code{\link[networkD3]{forceNetwork}}
 \code{\link[ggnetwork]{geom_edges}},\code{\link[ggnetwork]{geom_nodes}},\code{\link[ggnetwork]{geom_nodetext}},\code{\link[ggnetwork]{theme_blank}},\code{\link[ggnetwork]{geom_nodetext_repel}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes_}},
}
\author{
Fabricio Almeida-Silva
}
