% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{probe2gene}
\alias{probe2gene}
\title{Collapse probe-level expression matrix (or data frame) to gene-level}
\usage{
probe2gene(probe_exp, correspondence)
}
\arguments{
\item{probe_exp}{Expression matrix or data frame containing probe names as row names and chips as column names}

\item{correspondence}{A 2-column data frame containing gene IDs in the first column and their corresponding probe names in the second column. Probe names must be unique, but gene IDs can be repeated, since different probes can match the same gene.}
}
\value{
A data frame containing gene IDs as row names and arrays in the column names.
}
\description{
Collapse probe-level expression matrix (or data frame) to gene-level
}
\details{
If there are 2 or more probes for the same gene, the one with the highest mean will be selected.
}
\seealso{
\code{\link[WGCNA]{collapseRows}}
}
