% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{q_normalize}
\alias{q_normalize}
\title{Quantile normalize the expression data}
\usage{
q_normalize(exp)
}
\arguments{
\item{exp}{Gene expression matrix, where rownames are sample names and colnames are gene IDs.}
}
\value{
Expression matrix with normalized values
}
\description{
Quantile normalize the expression data
}
