% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{remove_na}
\alias{remove_na}
\title{Remove missing values in a gene expression data frame}
\usage{
remove_na(exp, replaceby = 0)
}
\arguments{
\item{exp}{Gene expression data frame or with genes in row names and samples IDs in column names.}

\item{replaceby}{What to use instead of NAs. One of 0 or 'mean'. Default is 0.}
}
\value{
Gene expression data frame with all NAs replaced according to the argument 'replaceby'
}
\description{
Remove missing values in a gene expression data frame
}
\author{
Fabricio Almeida-Silva
}
