% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se.seed}
\alias{se.seed}
\title{Soybean seed gene expression data from SRA study PRJNA197251}
\format{
An object of class \code{SummarizedExperiment}
}
\usage{
se.seed
}
\description{
Filtered expression data in transcripts per million (TPM) from Danzer et al., 2015.
Genes with TPM values < 2 in any of samples were removed to reduce package size.
The expression data and associated sample metadata are stored in a SummarizedExperiment
object with 10381 genes and 51 samples.
}
\examples{
data(se.seed)
}
\references{
Danzer, J., Mellott, E., Bui, A. Q., Le, B. H., Martin, P., Hashimoto, M., ... & Goldberg, R. B. (2015).
Down-regulating the expression of 53 soybean transcription factor genes uncovers a role for SPEECHLESS in
initiating stomatal cell lineages during embryo development. Plant Physiology, 168(3), 1025-1035.
}
\keyword{datasets}
