% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soybean_interpro}
\alias{soybean_interpro}
\title{Soybean Interpro annotation}
\format{
A 2-column data frame containing gene IDs and their associated Intepro annotations.
}
\usage{
soybean_interpro
}
\description{
Interpro protein domain annotation retrieved from the PLAZA 4.0 database.
Only genes included in \code{se.seed} are present in this subset.
}
\examples{
data(soybean_interpro)
}
\references{
Van Bel, M., Diels, T., Vancaester, E., Kreft, L., Botzki, A., Van de Peer, Y., ... & Vandepoele, K. (2018). PLAZA 4.0: an integrative resource for functional, evolutionary and comparative plant genomics. Nucleic acids research, 46(D1), D1190-D1196.
}
\keyword{datasets}
