% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{ZKfiltering}
\alias{ZKfiltering}
\title{Filter outlying samples based on standardized connectivity (Zk) method}
\usage{
ZKfiltering(raw_exp, Zk = -2, cor_method = "spearman")
}
\arguments{
\item{raw_exp}{Raw gene expression table, where rownames are gene IDs and colnames are sample names.}

\item{Zk}{Standardized connectivity threshold. Default is -2.}

\item{cor_method}{Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman", considering that the expression data does not follow a normal distribution.}
}
\value{
Filtered gene expression dataframe.
}
\description{
Filter outlying samples based on standardized connectivity (Zk) method
}
\seealso{
\code{\link[WGCNA]{adjacency}}
}
\author{
Fabricio Almeida-Silva
}
