% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zma.tfs}
\alias{zma.tfs}
\title{Maize transcription factors}
\format{
A data frame with gene IDs of TFs and their associated families.
}
\usage{
zma.tfs
}
\description{
Transcription factors and their families were downloaded from PlantTFDB 4.0.
}
\examples{
data(zma.tfs)
}
\references{
Jin, J., Tian, F., Yang, D. C., Meng, Y. Q., Kong, L., Luo, J., &
Gao, G. (2016). PlantTFDB 4.0: toward a central hub for transcription
factors and regulatory interactions in plants. Nucleic acids research, gkw982.
}
\keyword{datasets}
