% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_modules.R
\name{consensus_trait_cor}
\alias{consensus_trait_cor}
\title{Correlate set-specific modules and consensus modules to sample information}
\usage{
consensus_trait_cor(
  consensus,
  cor_method = "spearman",
  continuous_trait = FALSE,
  palette = "RdYlBu",
  cex.lab.x = 0.6,
  cex.lab.y = 0.6,
  cex.text = 0.6,
  save_sets = FALSE,
  transpose = FALSE
)
}
\arguments{
\item{consensus}{Consensus network returned by \code{consensus_modules}.}

\item{cor_method}{Correlation method to be used. One of 'spearman' or 'pearson'. Default is 'spearman'.}

\item{continuous_trait}{Logical indicating if trait is a continuous variable. Default is FALSE.}

\item{palette}{RColorBrewer's color palette to use. Default is "RdYlBu", a palette ranging from blue to red.}

\item{cex.lab.x}{Font size for x axis labels. Default: 0.6.}

\item{cex.lab.y}{Font size for y axis labels. Default: 0.6.}

\item{cex.text}{Font size for numbers inside matrix. Default: 0.6.}

\item{save_sets}{Logical indicating whether to save heatmaps of correlations between set-specific modules and traits as a PDF figure. Default: FALSE.}

\item{transpose}{Logical indicating whether to transpose the heatmap of not. Default is FALSE.}
}
\value{
Heatmaps of relationships between consensus/set-specific modules and sample information (metadata) in a PDF file.
}
\description{
Correlate set-specific modules and consensus modules to sample information
}
\details{
Significance levels:
1 asterisk: significant at alpha = 0.05.
2 asterisks: significant at alpha = 0.01.
3 asterisks: significant at alpha = 0.001.
no asterisk: not significant.
}
\examples{
set.seed(12)
data(zma.se)
filt.zma <- filter_by_variance(zma.se, n=500)
zma.set1 <- filt.zma[, sample(colnames(filt.zma), size=20, replace=FALSE)]
zma.set2 <- filt.zma[, sample(colnames(filt.zma), size=20, replace=FALSE)]
list.sets <- list(zma.set1, zma.set2)
# SFT power previously identified with consensus_SFT_fit()
consensus <- consensus_modules(list.sets, power = c(11, 13),
                               cor_method = "pearson")
consensus_trait <- consensus_trait_cor(consensus, cor_method = "pearson")
}
\seealso{
\code{\link[WGCNA]{corPvalueFisher}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{labeledHeatmap}},\code{\link[WGCNA]{blueWhiteRed}}
}
