% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Perform enrichment analysis for a set of genes}
\usage{
enrichment_analysis(
  genes,
  exp,
  annotation,
  column,
  correction = "BH",
  p = 0.05
)
}
\arguments{
\item{genes}{Character vector containing genes for overrepresentation analysis.}

\item{exp}{Gene expression data frames with gene IDs in row names and column names in column names.}

\item{annotation}{Annotation data frame with genes in the first column and functional annotation in the other columns, which can be exported from Biomart or similar databases.}

\item{column}{Column name on the data frame specified in 'annotation' used to test for enrichment (e.g. "GO.ID").}

\item{correction}{Multiple testing correction method. One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none". Default is "BH".}

\item{p}{P-value threshold. P-values below this threshold will be considered significant. Default is 0.05.}
}
\value{
List containing two elements: \describe{
 \item{Significant_terms}{Data frame contaning significant terms and p-values}
 \item{Associated_genes}{List containing genes associated with each significant term}
}
}
\description{
Perform enrichment analysis for functional annotation such as Gene Ontology (GO), pathway (KEGG, MapMan) or protein domains (PFAM, Panther).
}
\seealso{
\code{\link[bc3net]{enrichment}}
}
\author{
Fabricio Almeida-Silva
}
