% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Perform enrichment analysis for a set of genes}
\usage{
enrichment_analysis(
  genes,
  background_genes,
  annotation,
  column = NULL,
  correction = "BH",
  p = 0.05
)
}
\arguments{
\item{genes}{Character vector containing genes for overrepresentation analysis.}

\item{background_genes}{Character vector of genes to be used as background for the Fisher's Exact Test.}

\item{annotation}{Annotation data frame with genes in the first column and functional annotation in the other columns, which can be exported from Biomart or similar databases.}

\item{column}{Column or columns of  \code{annotation} to be used for enrichment. Both character or numeric values with column indices can be used. If users want to supply more than one column, input a character or numeric vector. Default: all columns from \code{annotation}.}

\item{correction}{Multiple testing correction method. One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none". Default is "BH".}

\item{p}{P-value threshold. P-values below this threshold will be considered significant. Default is 0.05.}
}
\value{
Data frame contaning significant terms, p-values and associated genes
}
\description{
Perform enrichment analysis for functional annotation such as Gene Ontology (GO), pathway (KEGG, MapMan) or protein domains (PFAM, Panther).
}
\author{
Fabricio Almeida-Silva
}
