% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{exp2gcn}
\alias{exp2gcn}
\title{Reconstruct gene coexpression network from gene expression}
\usage{
exp2gcn(
  exp,
  net_type = "signed",
  module_merging_threshold = 0.8,
  SFTpower = NULL,
  cor_method = "spearman",
  reportPDF = TRUE
)
}
\arguments{
\item{exp}{A gene expression data frame with genes in row names and samples in column names or a `SummarizedExperiment` object.}

\item{net_type}{Network type. One of 'signed', 'signed hybrid' or 'unsigned'. Default: 'signed'.}

\item{module_merging_threshold}{Correlation threshold to merge similar modules into a single one. Default: 0.8.}

\item{SFTpower}{SFT power generated by the function \code{SFT_fit}.}

\item{cor_method}{Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman", considering that the expression data does not follow a normal distribution.}

\item{reportPDF}{Logical indicating whether to save 2 PDF figures with heatmap of eigengene networks and dendrogram with modules, respectively. Default: TRUE.}
}
\value{
List containing: \itemize{
  \item Adjacency matrix
  \item Data frame of module eigengenes
  \item Data frame of genes and their corresponding modules
  \item Data frame of intramodular connectivity
  \item Vector of module assignment
  \item Correlation matrix
  \item Parameters used for network reconstruction
}
}
\description{
Reconstruct gene coexpression network from gene expression
}
\examples{
data(filt.se)
# The SFT fit was previously calculated and the optimal power was 16
gcn <- exp2gcn(filt.se, SFTpower = 16, cor_method = "pearson", reportPDF = FALSE)
}
\seealso{
\code{\link[WGCNA]{adjacency.fromSimilarity}},\code{\link[WGCNA]{TOMsimilarity}},\code{\link[WGCNA]{standardColors}},\code{\link[WGCNA]{labels2colors}},\code{\link[WGCNA]{moduleEigengenes}},\code{\link[WGCNA]{plotEigengeneNetworks}},\code{\link[WGCNA]{mergeCloseModules}},\code{\link[WGCNA]{plotDendroAndColors}},\code{\link[WGCNA]{intramodularConnectivity}}
 \code{\link[dynamicTreeCut]{cutreeDynamicTree}}
}
\author{
Fabricio Almeida-Silva
}
