% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcn_inference.R
\name{gene_significance}
\alias{gene_significance}
\title{Calculate gene significance for a given group of genes}
\usage{
gene_significance(
  exp,
  metadata,
  alpha = 0.05,
  min_cor = 0,
  use_abs = TRUE,
  savetofile = FALSE,
  palette = "RdYlBu",
  show_rownames = FALSE,
  continuous_trait = FALSE
)
}
\arguments{
\item{exp}{Data frame of gene expression with gene IDs in row names and sample names in column names.}

\item{metadata}{Data frame containing sample names on the first column and sample information on the second column.}

\item{alpha}{Significance level. Default is 0.05.}

\item{min_cor}{Minimum correlation coefficient. Default is 0, which excludes negative correlations.}

\item{use_abs}{Logical indicating whether to filter by correlation using absolute value or not. If TRUE, a \code{min_cor} of say 0.4 would keep all correlations above 0.4 and below -0.4. Default is TRUE.}

\item{savetofile}{Logical indicating whether to save the table to correlations and p-values to a tab-delimited file or not.}

\item{palette}{RColorBrewer's color palette to use. Default is "RdYlBu", a palette ranging from blue to red.}

\item{show_rownames}{Logical indicating whether to show row names or not. Default is FALSE.}

\item{continuous_trait}{Logical indicating if trait is a continuous variable. Default is FALSE.}
}
\value{
A heatmap of gene significance (GS) and a list containing filtered and raw GS.
}
\description{
Calculate gene significance for a given group of genes
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[WGCNA]{corPvalueStudent}}
 \code{\link[pheatmap]{pheatmap}}
 \code{\link[RColorBrewer]{RColorBrewer}}
}
