% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{get_HK}
\alias{get_HK}
\title{Get a vector object with housekeeping genes}
\usage{
get_HK(exp)
}
\arguments{
\item{exp}{Expression dataframe, where rownames are gene IDs and colnames are sample names}
}
\value{
vector object of gene IDs of housekeeping genes
}
\description{
Get a vector object with housekeeping genes
}
\author{
Fabricio Almeida-Silva
}
