% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{get_hubs}
\alias{get_hubs}
\title{Get hub genes in a given module or in all modules at once}
\usage{
get_hubs(exp, genes_modules, MEs, kIN, cor_method = "spearman")
}
\arguments{
\item{exp}{Gene expression data frame with gene IDs as row names and sample names as column names.}

\item{genes_modules}{A 2-column data frame containing genes on the first column and modules on the second column. It is the 3rd element of the output from \code{exp2net}.}

\item{MEs}{Data frame of module eigengenes. It is the 2nd element of the output from \code{exp2net}.}

\item{kIN}{Data frame with intramodular connectivity (kIN) for each gene. It is the 4th element of the output from \code{exp2net}.}

\item{cor_method}{Correlation method. One of 'pearson' or 'spearman' (recommended). Default is 'spearman'.}
}
\value{
Data frame containing gene IDs, modules and intramodular connectivity of all hubs.
}
\description{
Get hub genes in a given module or in all modules at once
}
\seealso{
\code{\link[WGCNA]{signedKME}}
}
\author{
Fabricio Almeida-Silva
}
