% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_inference.R
\name{grn_clr}
\alias{grn_clr}
\title{Infer gene regulatory network with the Context Likelihood of Relatedness (CLR) algorithm}
\usage{
grn_clr(exp, estimator = "pearson")
}
\arguments{
\item{exp}{Expression matrix with gene IDs as row names and samples as column names.}

\item{estimator}{Entropy estimator to be used. One of "mi.empirical", "mi.mm", "mi.shrink", "mi.sg", "pearson", "spearman", or "kendall". Default: "pearson".}
}
\value{
A gene regulatory network represented as an edge list.
}
\description{
Infer gene regulatory network with the Context Likelihood of Relatedness (CLR) algorithm
}
