% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_inference.R
\name{grn_combined}
\alias{grn_combined}
\title{Infer gene regulatory network with multiple algorithms and combine results in a list}
\usage{
grn_combined(exp, regulators = NULL, ...)
}
\arguments{
\item{exp}{Expression matrix with gene IDs as row names and samples as column names.}

\item{regulators}{A character vector of regulators (e.g., transcription factors or miRNAs). All regulators must be included in `exp`.}

\item{...}{Additional arguments passed to `GENIE3::GENIE3()`.}
}
\value{
A list of data frames representing edge lists. Each list element is an edge list for a specific method.
}
\description{
Infer gene regulatory network with multiple algorithms and combine results in a list
}
