% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{modPres_WGCNA}
\alias{modPres_WGCNA}
\title{Calculate module preservation between two expression data sets using WGCNA's algorithm}
\usage{
modPres_WGCNA(
  explist,
  ref_net,
  savePreservation = TRUE,
  plot_all_stats = FALSE,
  calculateClusterCoeff = FALSE
)
}
\arguments{
\item{explist}{List of expression matrices with gene/probe names corresponding to column names and sample names corresponding to row names.}

\item{ref_net}{Reference network object returned by the function \code{exp2net}.}

\item{savePreservation}{Logical indicating whether to save module preservation into an R object or not. As the calculation of module preservation can take a long time, it is useful to save time in future analyses. Default is TRUE.}

\item{plot_all_stats}{Logical indicating whether to save all density and connectivity statistics in a PDF file or not. Default is FALSE.}

\item{calculateClusterCoeff}{Logical indicating if clustering coefficient statistics should be calculated. Only valid if \code{plot_all_stats} is TRUE. As it may take a long time, default is FALSE.}
}
\value{
A list object resulting from \code{WGCNA::modulePreservation} and .pdf files with plot statistics.
}
\description{
Calculate module preservation between two expression data sets using WGCNA's algorithm
}
\seealso{
\code{\link[WGCNA]{modulePreservation}},\code{\link[WGCNA]{standardColors}}
 \code{\link[ggpubr]{ggscatter}},\code{\link[ggpubr]{ggarrange}},\code{\link[ggpubr]{ggexport}}
 \code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{geom_abline}}
}
