% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{module_preservation}
\alias{module_preservation}
\title{Calculate module preservation between two expression data sets}
\usage{
module_preservation(
  exp_ref,
  exp_test,
  moduleColors,
  net_type = "signed",
  cor_method = "spearman",
  savePreservation = TRUE,
  plot_all_stats = FALSE,
  calculateClusterCoeff = FALSE
)
}
\arguments{
\item{exp_ref}{Data frame of reference gene expression data set, with gene IDs in row names and sample names in column names.}

\item{exp_test}{Data frame of test gene expression data set, with gene IDs in row names and sample names in column names.}

\item{moduleColors}{Vector of module assignment returned by \code{exp2net}.}

\item{net_type}{Network type. One of 'signed', 'signed hybrid' or 'unsigned'. Default is signed.}

\item{cor_method}{Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman".}

\item{savePreservation}{Logical indicating whether to save module preservation into an R object or not. As the calculation of module preservation can take a long time, it is useful to save time in future analyses. Default is TRUE.}

\item{plot_all_stats}{Logical indicating whether to save all density and connectivity statistics in a PDF file or not. Default is FALSE.}

\item{calculateClusterCoeff}{Logical indicating if clustering coefficient statistics should be calculated. Only valid if \code{plot_all_stats} is TRUE. As it may take a long time, default is FALSE.}
}
\value{
A list object resulting from \code{WGCNA::modulePreservation} and .pdf files with plot statistics.
}
\description{
Calculate module preservation between two expression data sets
}
\seealso{
\code{\link[WGCNA]{modulePreservation}},\code{\link[WGCNA]{standardColors}}
 \code{\link[ggpubr]{ggscatter}},\code{\link[ggpubr]{ggarrange}},\code{\link[ggpubr]{ggexport}}
 \code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{geom_abline}}
}
