% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{module_preservation}
\alias{module_preservation}
\title{Calculate network preservation between two expression data sets}
\usage{
module_preservation(
  explist,
  ref_net = NULL,
  test_net = NULL,
  algorithm = "netrep",
  diffIDs = FALSE,
  correspondence = NULL,
  nPerm = 1000,
  nThreads = 2,
  savePreservation = TRUE,
  plot_all_stats = FALSE,
  calculateClusterCoeff = FALSE
)
}
\arguments{
\item{explist}{List of expression matrices with gene/probe names corresponding to column names and sample names corresponding to row names.}

\item{ref_net}{Reference network object returned by the function \code{exp2net}.}

\item{test_net}{Test network object returned by the function \code{exp2net}.}

\item{algorithm}{Module preservation algorithm to be used. One of 'netrep' (default, permutation-based) or WGCNA.}

\item{diffIDs}{Logical indicating whether the different data sets have different gene IDs (e.g., for different species). Default: FALSE}

\item{correspondence}{Data frame containing gene IDs for the reference data set in the column 1 and gene IDs for the test data set in column 2. Only required if diffIDs is TRUE.}

\item{nPerm}{Number of permutations. Default: 1000}

\item{nThreads}{Number of threads to be used for parallel computing. Default: 2}

\item{savePreservation}{Logical indicating whether to save module preservation into an R object or not. As the calculation of module preservation can take a long time, it is useful to save time in future analyses. Default is TRUE.}

\item{plot_all_stats}{Logical indicating whether to save all density and connectivity statistics in a PDF file or not. Default is FALSE.}

\item{calculateClusterCoeff}{Logical indicating if clustering coefficient statistics should be calculated. Only valid if \code{plot_all_stats} is TRUE. As it may take a long time, default is FALSE.}
}
\value{
A list containing the preservation statistics. The list will be different depending on the chosen algorithm. See \code{WGCNA::modulePreservation} or \code{NetRep::modulePreservation} for more info.
}
\description{
Calculate network preservation between two expression data sets
}
