% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{module_stability}
\alias{module_stability}
\title{Perform module stability analysis}
\usage{
module_stability(
  norm.exp,
  moduleColors,
  net_type = "signed hybrid",
  SFTpower,
  cor_method = "bicor",
  module_merging_threshold = 0.2,
  nRuns = 30
)
}
\arguments{
\item{norm.exp}{Normalized gene expression data frame, where row names are gene IDs and column names are sample names.}

\item{moduleColors}{Vector of module assignment returned by \code{exp2net}.}

\item{net_type}{Network type. One of 'signed', 'signed hybrid' or 'unsigned'. Default is "signed hybrid".}

\item{SFTpower}{SFT power generated by the function \code{SFT_fit}.}

\item{cor_method}{One of "bicor" (default) or "pearson". Spearman correlation coefficients are not allowed for this analysis.}

\item{module_merging_threshold}{Maximum dissimilarity (1-cor) between module for module merging. As it is based on dissimilarity, a threshold of 0.2 means that module eigengenes must be at least 0.8 similar to be merged. Default is 0.2.}

\item{nRuns}{Number of times to resample. Default is 30.}
}
\value{
A .pdf file in the current working directory representing the module stability across the resamplings.
}
\description{
Perform module stability analysis
}
\seealso{
\code{\link[WGCNA]{sampledBlockwiseModules}},\code{\link[WGCNA]{Inline display of progress}},\code{\link[WGCNA]{matchLabels}},\code{\link[WGCNA]{plotDendroAndColors}}
}
