% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{module_trait_cor}
\alias{module_trait_cor}
\title{Correlate module eigengenes to trait}
\usage{
module_trait_cor(
  exp,
  metadata,
  MEs,
  cor_method = "spearman",
  transpose = FALSE,
  palette = "RdYlBu",
  continuous_trait = FALSE,
  cex.lab.x = 0.6,
  cex.lab.y = 0.6,
  cex.text = 0.6
)
}
\arguments{
\item{exp}{Gene expression data frame used to reconstruct the network, with gene IDs as row names and sample names as column names.}

\item{metadata}{A 2-column trait data frame containing sample names in the first column and sample descriptions in the second column.}

\item{MEs}{Module eigengenes. It is the 2nd element of the result list generated by the function \code{exp2net}.}

\item{cor_method}{Method to calculate correlation. One of 'pearson', 'spearman' or 'kendall'. Default is 'spearman'.}

\item{transpose}{Logical indicating whether to transpose the heatmap of not. Default is FALSE.}

\item{palette}{RColorBrewer's color palette to use. Default is "RdYlBu", a palette ranging from blue to red.}

\item{continuous_trait}{Logical indicating if trait is a continuous variable. Default is FALSE.}

\item{cex.lab.x}{Font size for x axis labels. Default: 0.6.}

\item{cex.lab.y}{Font size for y axis labels. Default: 0.6.}

\item{cex.text}{Font size for numbers inside matrix. Default: 0.6.}
}
\value{
A heatmap showing the correlations between module eigengenes and trait with their associated significance levels.
}
\description{
Correlate module eigengenes to trait
}
\details{
Significance levels:
1 asterisk: significant at alpha = 0.05.
2 asterisks: significant at alpha = 0.01.
3 asterisks: significant at alpha = 0.001.
no asterisk: not significant.
}
\seealso{
\code{\link[WGCNA]{corPvalueStudent}},\code{\link[WGCNA]{labeledHeatmap}},\code{\link[WGCNA]{blueWhiteRed}}
}
\author{
Fabricio Almeida-Silva
}
