% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_expression_profile}
\alias{plot_expression_profile}
\title{Plot expression profile of given genes across samples}
\usage{
plot_expression_profile(
  genes,
  exp,
  metadata,
  plot_module = TRUE,
  genes_modules,
  modulename
)
}
\arguments{
\item{genes}{Character vector containing a subset of genes from which edges will be extracted. It can be ignored if \code{plot_module} is TRUE.}

\item{exp}{Data frame containing genes IDs in row names and sample names in column names.}

\item{metadata}{A 2-column data frame containing sample names in the first column and sample descriptions in the second column.}

\item{plot_module}{Logical indicating whether to plot a whole module or not. If set to FALSE, \code{genes} must be specified.}

\item{genes_modules}{Data frame containing genes in column 1 and their corresponding modules in column 2. It is the third element of the output list from \code{exp2net}.}

\item{modulename}{Character with name of the module to be plotted. To include 2 or more modules, input the names in a character vector.}
}
\value{
A ggplot object showing the expression profile of some genes across all samples.
}
\description{
Plot expression profile of given genes across samples
}
\seealso{
\code{\link[reshape2]{melt}}
}
\author{
Fabricio Almeida-Silva
}
