% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_expression_profile}
\alias{plot_expression_profile}
\title{Plot expression profile of given genes across samples}
\usage{
plot_expression_profile(
  genes,
  exp,
  metadata,
  plot_module = TRUE,
  net,
  modulename
)
}
\arguments{
\item{genes}{Character vector containing a subset of genes from which edges will be extracted. It can be ignored if \code{plot_module} is TRUE.}

\item{exp}{A gene expression data frame with genes in row names and samples in column names or a `SummarizedExperiment` object.}

\item{metadata}{A data frame containing sample names in row names and sample annotation in the first column. Ignored if `exp` is a `SummarizedExperiment` object, since the function will extract colData.}

\item{plot_module}{Logical indicating whether to plot a whole module or not. If set to FALSE, \code{genes} must be specified.}

\item{net}{List object returned by \code{exp2net}.}

\item{modulename}{Name of the module to plot.}
}
\value{
A ggplot object showing the expression profile of some genes across all samples.
}
\description{
Plot expression profile of given genes across samples
}
\examples{
data(se.seed)
data(filt.se)
genes <- rownames(filt.se)
plot_expression_profile(genes=genes, exp=se.seed, plot_module=FALSE)
}
\seealso{
\code{\link[reshape2]{melt}}
}
\author{
Fabricio Almeida-Silva
}
