% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{plot_gcn}
\alias{plot_gcn}
\title{Plot gene coexpression network from edge list}
\usage{
plot_gcn(
  edgelist_gcn,
  genes_modules = NULL,
  modulename = NULL,
  kIN = NULL,
  hubs = NULL,
  show_labels = "tophubs",
  top_n_hubs = 5,
  interactive = FALSE
)
}
\arguments{
\item{edgelist_gcn}{Data frame containing the edge list for the PPI or GRN network. First column is the protein 1 (or TF) and second column is the protein 2 (or target gene). All other columns are interpreted as edge attributes.}

\item{genes_modules}{Data frame containing genes in column 1 and their corresponding modules in column 2. It is the third element of the output list from \code{exp2net}.}

\item{modulename}{Character with the name of the module to be plotted.}

\item{kIN}{Data frame containing intramodular connectivity for each gene in the network. It is the fourth element of the output list from \code{exp2net}.}

\item{hubs}{Data frame containing hub genes in the first column, their modules in the second column, and intramodular connectivity in the third column.}

\item{show_labels}{Character indicating which nodes will be labeled. One of "all", "allhubs", "tophubs", or "none".}

\item{top_n_hubs}{Number of top hubs to be labeled. It is only valid if \code{show_labels} equals "tophubs". Default is 5.}

\item{interactive}{Logical indicating whether the network should be interactive or not. Default is FALSE.}
}
\description{
Plot gene coexpression network from edge list
}
\seealso{
\code{\link[igraph]{simplify}},\code{\link[igraph]{as_data_frame}},\code{\link[igraph]{gorder}}
 \code{\link[networkD3]{igraph_to_networkD3}},\code{\link[networkD3]{forceNetwork}}
 \code{\link[ggnetwork]{c("ggnetwork", "ggnetwork", "ggnetwork")}},\code{\link[ggnetwork]{geom_edges}},\code{\link[ggnetwork]{geom_nodes}},\code{\link[ggnetwork]{geom_nodetext}},\code{\link[ggnetwork]{theme_blank}},\code{\link[ggnetwork]{geom_nodetext_repel}},\code{\link[ggnetwork]{c("unit", "unit")}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}}
}
\author{
Fabricio Almeida-Silva
}
