% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_gcn}
\alias{plot_gcn}
\title{Plot gene coexpression network from edge list}
\usage{
plot_gcn(
  edgelist_gcn,
  net,
  color_by = "module",
  hubs = NULL,
  show_labels = "tophubs",
  top_n_hubs = 5,
  interactive = FALSE
)
}
\arguments{
\item{edgelist_gcn}{Data frame containing the edge list for the GCN.
The edge list can be generated with \code{get_edge_list()}.}

\item{net}{List object returned by \code{exp2net}.}

\item{color_by}{How should nodes be colored? It must be either "module"
(nodes will have the colors of their modules) or a 2-column data frame
containing genes in the first column and a custom gene annotation
in the second column. Default: "module".}

\item{hubs}{Data frame containing hub genes in the first column,
their modules in the second column, and intramodular connectivity in
the third column.}

\item{show_labels}{Character indicating which nodes will be labeled.
One of "all", "allhubs", "tophubs", or "none". Default: tophubs.}

\item{top_n_hubs}{Number of top hubs to be labeled. It is only valid
if \code{show_labels} equals "tophubs". Default is 5.}

\item{interactive}{Logical indicating whether the network should be
interactive or not. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Plot gene coexpression network from edge list
}
\examples{
data(filt.se)
gcn <- exp2gcn(filt.se, SFTpower = 18, cor_method = "pearson",
               reportPDF = FALSE)
gcn_edges <- get_edge_list(gcn, module="brown", filter=TRUE,
                           method="min_cor")
hubs <- get_hubs_gcn(filt.se, gcn)
p <- plot_gcn(gcn_edges, gcn, hubs = hubs)
}
\seealso{
\code{\link[igraph]{simplify}},\code{\link[igraph]{as_data_frame}},\code{\link[igraph]{gorder}}
 \code{\link[networkD3]{igraph_to_networkD3}},\code{\link[networkD3]{forceNetwork}}
 \code{\link[ggnetwork]{geom_edges}},\code{\link[ggnetwork]{geom_nodes}},\code{\link[ggnetwork]{geom_nodetext}},\code{\link[ggnetwork]{theme_blank}},\code{\link[ggnetwork]{geom_nodetext_repel}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes_}}
}
\author{
Fabricio Almeida-Silva
}
