% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap of hierarchically clustered sample correlations or gene expression}
\usage{
plot_heatmap(
  exp,
  col_metadata = NULL,
  row_metadata = NULL,
  cor_method = "spearman",
  type = "samplecor",
  palette = NULL,
  log_trans = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  show_rownames = FALSE,
  show_colnames = TRUE,
  scale = "none",
  fontsize = 10,
  cutree_rows = 1,
  cutree_cols = 1
)
}
\arguments{
\item{exp}{Expression data frame with genes in rows and samples in columns.}

\item{col_metadata}{A data frame containing sample names in the first column and sample description in the subsequent columns. If there are more than 2 columns, additional columns will be interpreted as multiple column annotations.}

\item{row_metadata}{A data frame containing gene IDs in the first column and gene functional classification in the subsequent columns. If there are more than 2 columns, additional columns will be interpreted as multiple gene annotation classifications.}

\item{cor_method}{Correlation method. One of 'spearman' or 'pearson'. Default is 'spearman'.}

\item{type}{Type of heatmap to plot. One of 'samplecor' (sample correlations) or 'expr'. Default is 'samplecor'.}

\item{palette}{RColorBrewer palette to use. Default is "Blues" for sample correlation heatmap and "YlOrRd" for gene expression heatmap.}

\item{log_trans}{Logical. It specifies whether to log transform data or not. Default is FALSE.}

\item{cluster_rows}{Logical indicating whether to cluster rows or not. Default is TRUE.}

\item{cluster_cols}{Logical indicating whether to cluster columns or not. Default is TRUE.}

\item{show_rownames}{Logical indicating whether to show row names or not. Default is FALSE.}

\item{show_colnames}{Logical indicating whether to show column names or not. Default is TRUE.}

\item{scale}{Character indicating if values should be centered and scaled in rows, columns, or none. One of 'row', 'column', or 'none'. Default is 'none'.}

\item{fontsize}{Base fontsize for the plot.}

\item{cutree_rows}{Number of clusters into which rows are divided. Default is 1, which is equivalent to no division.}

\item{cutree_cols}{Number of clusters into which columns are divided. Default is 1, which is equivalent to no division.}
}
\value{
heatmap of hierarchically clustered samples with metadata information (optional)
}
\description{
Plot heatmap of hierarchically clustered sample correlations or gene expression
}
\seealso{
\code{\link[RColorBrewer]{RColorBrewer}}
 \code{\link[pheatmap]{pheatmap}}
}
\author{
Fabricio Almeida-Silva
}
