% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_ngenes_per_module}
\alias{plot_ngenes_per_module}
\title{Plot number of genes per module}
\usage{
plot_ngenes_per_module(net = NULL)
}
\arguments{
\item{net}{List object returned by \code{exp2gcn}.}
}
\value{
A ggplot object with a bar plot of gene number in each module.
}
\description{
Plot number of genes per module
}
\examples{
data(filt.se)
gcn <- exp2gcn(filt.se, SFTpower = 16, cor_method = "pearson", reportPDF = FALSE)
plot_ngenes_per_module(gcn)
}
\seealso{
\code{\link[ggpubr]{ggbarplot}}
}
