% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot Principal Component Analysis (PCA) of samples}
\usage{
plot_PCA(
  exp,
  metadata,
  log_trans = FALSE,
  PCs = "1x2",
  size = 2,
  interactive = FALSE
)
}
\arguments{
\item{exp}{Expression data frame with genes in rownames and samples in column names.}

\item{metadata}{Data frame containing sample IDs in first column and sample description (e.g. tissue) in the second column.}

\item{log_trans}{Logical. If TRUE, the expression data frame will be log transformed by log2(exp+1).}

\item{PCs}{Principal Components to be plotted on the x-axis and y-axis, respectively. One of "1x2", "1x3" or "2x3. Default is "1x2".}

\item{size}{Numeric indicating the point size. Default is 2.}

\item{interactive}{Logical indicating whether PCA should be plotted using Shiny's interactivity on RStudio. Default is FALSE.}
}
\value{
A ggplot object with the PCA plot.
}
\description{
Plot Principal Component Analysis (PCA) of samples
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Fabricio Almeida-Silva
}
