% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{q_normalize}
\alias{q_normalize}
\title{Quantile normalize the expression data}
\usage{
q_normalize(exp)
}
\arguments{
\item{exp}{A gene expression data frame with genes in row names
and samples in column names.}
}
\value{
Expression matrix with normalized values
}
\description{
Quantile normalize the expression data
}
\examples{
data(zma.se)
exp <- SummarizedExperiment::assay(zma.se)
norm_exp <- q_normalize(exp)
}
