% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction_and_analysis.R
\name{SFT_fit}
\alias{SFT_fit}
\title{Pick power to fit network to a scale-free topology}
\usage{
SFT_fit(exp, net_type = "signed", rsquared = 0.8, cor_method = "spearman")
}
\arguments{
\item{exp}{Normalized expression table, where rownames are gene IDs and colnames are sample names}

\item{net_type}{Network type. One of 'signed', 'signed hybrid' or 'unsigned'. Default is signed.}

\item{rsquared}{R squared cutoff. Default is 0.8.}

\item{cor_method}{Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman".}
}
\value{
Power to fit network to a scale-free topology and SFT fit plots in PDF in the user's working directory
}
\description{
Pick power to fit network to a scale-free topology
}
\seealso{
\code{\link[WGCNA]{pickSoftThreshold}}
}
\author{
Fabricio Almeida-Silva
}
